move {Duration,Float}Range params; godocs.
